+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.60.10] - 2020-05-12
+- <no-jira> - Fix to Busines Hours rule when clicking on context menu to perform an action
+- <no-jira> - Fix to DatePicker component that would show the DayPicker smaller than it should be
+
 ## [0.60.9] - 2020-05-07
 - <no-ticket> - Fix to Business Hours Rule component when used as a redux-form field